DEV-2234: Make TMB and TML optional in molecular record

+++ common/src/test/java/com/hartwig/actin/molecular/orange/interpretation/OrangeInterpreterTest.java
@@ -33,7 +33,7 @@ public class OrangeInterpreterTest {
         assertFalse(record.isMicrosatelliteUnstable());
         assertFalse(record.isHomologousRepairDeficient());
         assertEquals(8D, record.tumorMutationalBurden(), EPSILON);
-        assertEquals(100, record.tumorMutationalLoad());
+        assertEquals(100, (int) record.tumorMutationalLoad());
 
         MolecularEvidence evidence = record.evidence();
